Minor styling changes to the search page
