+++ iosApp/iosApp/ImageResultViewEntity+Extensions.swift
@@ -1,0 +1,7 @@
+
+import Foundation
+import ComposeApp
+
+extension ImageResultViewEntity : Identifiable, ObservableObject {
+    
+}
