+++ iosApp/iosApp/Scenes/Search/SearchResultsView.swift
@@ -1,0 +1,27 @@
+import Foundation
+import SwiftUI
+import ComposeApp
+
+struct SearchResultsView : View {
+    var results: [ImageResultViewEntity]
+    
+    let columns = [
+         GridItem(.adaptive(minimum: 80))
+     ]
+    
+    var body: some View {
+        ScrollView {
+            LazyVStack {
+                LazyVGrid(columns: columns, spacing: 20) {
+                                    ForEach(results.indices, id: \.self) { index in
+                                        ImageView(imageUrl: results[index].photo.url)
+                                            .frame(width: 100, height: 100)
+                                            .cornerRadius(8)
+                                    }
+                            }
+                            .padding(.horizontal)
+
+            }
+        }
+    }
+}
